build: create and own /var/lib/authselect
This directory will hold authselect state information such as copies
of currently generated files. These copies will be then used to test
whether the existing configuration was modified outside authselect
or not.

This approach allows us to introduce new command apply-changes that
will apply changes in profile without the need of using --force.
Otherwise when compared to profiles directly (current state), --force
parameter is required because the generated files seems to be changed.